Support switching the preview in studio mode
`scene switch` always changes the program, but I expect it to change the
preview in case the studio mode is enabled.

So change `scene switch` to switch the preview (if studio mode is
enabled) or the program (if studio mode is disabled). Add `scene
current` for always changing the program scene and add `scene preview`
to always change the preview scene (which requires studio mode to be
enabled).

Signed-off-by: Benjamin Drung <bdrung@posteo.de>